+++ src/bloqade/noise/native/stmts.py
@@ -4,7 +4,7 @@ from kirin import ir, types, lowering
 from kirin.decl import info, statement
 from kirin.dialects import ilist
 
-from bloqade.qasm2.types import QubitType
+from bloqade.types import QubitType
 
 from ._dialect import dialect
