+++ src/bloqade/qasm2/emit/gate.py
@@ -1,3 +1,4 @@
+from typing import Any
 from dataclasses import field, dataclass
 
 from kirin import ir, types, interp
@@ -6,6 +7,8 @@ from kirin.ir.dialect import Dialect as Dialect
 
 from bloqade.types import QubitType
 from bloqade.qasm2.parse import ast
+
+from bloqade.noise import native
 
 from .base import EmitError, EmitQASM2Base, EmitQASM2Frame
 
@@ -85,4 +88,85 @@ class Func(interp.MethodTable):
     @interp.impl(func.Return)
     @interp.impl(func.ConstantNone)
     def ignore(self, emit: EmitQASM2Gate, frame: EmitQASM2Frame, stmt):
+        return ()
+
+
+@native.dialect.register(key="emit.qasm2.gate")
+class NativeNoise(interp.MethodTable):
+
+    def _convert(self, node: ast.Bit | ast.Name) -> str:
+        if isinstance(node, ast.Bit):
+            return f"{node.name.id}[{node.addr}]"
+        else:
+            return f"{node.id}"
+
+    @interp.impl(native.CZPauliChannel)
+    def emit_czp(
+        self,
+        emit: EmitQASM2Gate,
+        frame: EmitQASM2Frame,
+        stmt: native.CZPauliChannel,
+    ):
+        paired: bool = stmt.paired
+        px_ctrl: float = stmt.px_ctrl
+        py_ctrl: float = stmt.py_ctrl
+        pz_ctrl: float = stmt.pz_ctrl
+        px_qarg: float = stmt.pz_qarg
+        py_qarg: float = stmt.py_qarg
+        pz_qarg: float = stmt.pz_qarg
+        ctrls: ilist.IList[ast.Bit, Any] = frame.get(stmt.ctrls)
+        qargs: ilist.IList[ast.Bit, Any] = frame.get(stmt.qargs)
+        frame.body.append(
+            ast.Comment(
+                text=f"native.CZPauliChannel(paired={paired}, p_ctrl[{px_ctrl}, {py_ctrl}, {pz_ctrl}], p_qarg[{px_qarg}, {py_qarg}, {pz_qarg}])"
+            )
+        )
+        frame.body.append(
+            ast.Comment(
+                text=f" -: ctrls: {', '.join([self._convert(q) for q in ctrls])}"
+            )
+        )
+        frame.body.append(
+            ast.Comment(
+                text=f" -: qargs: {', '.join([self._convert(q) for q in qargs])}"
+            )
+        )
+        return ()
+
+    @interp.impl(native.AtomLossChannel)
+    def emit_loss(
+        self,
+        emit: EmitQASM2Gate,
+        frame: EmitQASM2Frame,
+        stmt: native.AtomLossChannel,
+    ):
+        prob: float = stmt.prob
+        qargs: ilist.IList[ast.Bit, Any] = frame.get(stmt.qargs)
+        frame.body.append(ast.Comment(text=f"native.Atomloss(p={prob})"))
+        frame.body.append(
+            ast.Comment(
+                text=f" -: qargs: {', '.join([self._convert(q) for q in qargs])}"
+            )
+        )
+        return ()
+
+    @interp.impl(native.PauliChannel)
+    def emit_pauli(
+        self,
+        emit: EmitQASM2Gate,
+        frame: EmitQASM2Frame,
+        stmt: native.PauliChannel,
+    ):
+        px: float = stmt.px
+        py: float = stmt.py
+        pz: float = stmt.pz
+        qargs: ilist.IList[ast.Bit, Any] = frame.get(stmt.qargs)
+        frame.body.append(
+            ast.Comment(text=f"native.Atomloss(px={px}, py={py}, pz={pz})")
+        )
+        frame.body.append(
+            ast.Comment(
+                text=f" -: qargs: {', '.join([self._convert(q) for q in qargs])}"
+            )
+        )
         return ()
